Use unsaved UserDetails model in the bot
Not fixing name errors having renamed model.User to model._OLD_DEPRECATED_User because nothing should hit any code path
that tries to use it. This code is kept around solely as a reference; the purpose of this branch is to move useful parts
of this code into the new database-backed models.